refactor: 레거시 핸들러 및 매핑 시스템 제거
- 레거시 `Controller` 및 관련 클래스 제거
- 기존 핸들러 어댑터와 매핑 로직 재구성
- 패키지 리팩토링 및 불필요한 의존성 정리
- 어노테이션 기반 핸들러 매핑으로 전환
