--- mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/handler/adapter/ManualHandlerAdapter.java
@@ -1,26 +1,0 @@
-package com.interface21.webmvc.servlet.mvc.tobe.handler.adapter;
-
-import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerAdapter;
-import com.interface21.webmvc.servlet.view.JspView;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-public class ManualHandlerAdapter implements HandlerAdapter {
-
-    @Override
-    public boolean supports(final Object handler) {
-        return handler instanceof Controller;
-    }
-
-    @Override
-    public ModelAndView handle(
-            final HttpServletRequest request,
-            final HttpServletResponse response,
-            final Object handler
-    ) throws Exception {
-        var viewName = ((Controller) handler).execute(request, response);
-        return new ModelAndView(new JspView(viewName));
-    }
-}
